requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.50
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
